fix rendering of area dialog in hidpi displays
Renders the screenshot properly sized onto a hidpi device. Before this
fix the screenshot was rendered magnified by the device pixel ratio
because it assumed the screenshot is non-hdpi and needs therefore
scaling up.

By setting the screenshot's dpr to the same as the QPaintDevice's (QDialog)
we tell QT to not scale the image.